hz/requirements.txt
hazelcast-python-client==5.3.0
python-dotenv==1.0.1